Permit to hide connected user in activity area screen (#9381)
